Lock interfaces properties with different availability batch editing
When interfaces are edited in bulk for multiple nodes they might support
different set of properties/functions (e.g. dpdk or sriov). Current
behavior presumes these sets are similar for all nodes/interfaces and
no additional validation is performed. Due to this it is possible now
to modify properties that interface does not support.

This fix compares interfaces properties across the nodes and determines
intersected properties which can be modified in bulk. Other properties
get locked for editing and rendered with the lock icon.

Methods for changes detection and application are affected also not
to respect properties with different availability.

Change-Id: I38b0c8e73351dc85757f418ae413c9356b1c4936
Closes-Bug: #1561458